심볼 레벨 메소 BIGINT 확장
그랜드 어센틱심볼 상위 레벨 메소가 INT(2.1B) 범위를 초과하므로
SymbolLevel.meso_cost를 BIGINT로 변경.

+++ backend/models/symbol/SymbolLevel.js
@@ -6,7 +6,7 @@ export const SymbolLevel = sequelize.define('SymbolLevel', {
   symbol_id: { type: DataTypes.INTEGER, allowNull: false },
   level: { type: DataTypes.TINYINT, allowNull: false },
   required_count: { type: DataTypes.SMALLINT, allowNull: false },
-  meso_cost: { type: DataTypes.INTEGER, allowNull: false },
+  meso_cost: { type: DataTypes.BIGINT, allowNull: false },
 }, {
   tableName: 'sym_levels',
   underscored: true,
